feat: user avatar component
